+++ lib/moon/light/live_component.ex
@@ -1,13 +1,13 @@
-defmodule Moon.StatefulComponent do
+defmodule Moon.Light.LiveComponent do
   @moduledoc false
 
   defmacro __using__(opts \\ []) do
     quote do
-      use Surface.LiveComponent, unquote(opts)
+      use Phoenix.LiveComponent, unquote(opts)
       import Moon.Helpers.Styled
       import Moon.Helpers.MergeClass
       import Moon.Helpers.MoonRender
-      import Moon.Light
+      # import Moon.Light
 
       def __moduledoc__(), do: @moduledoc
     end